[Simulate] Change simulation, get sickness use norm
Change the get_sickness function to use norm(dot/max_dot-limit,std)*increase
as discussed in the meeting 17th of oct.
The limit, std and increase (sp, sa and si) are adjustable parameters.
When generating multiple datasets a list of different values for these parameters is used to create datasets with different characteristics