Evaluation framework - version 1
Tightly coupled and not very flexible.

The Evaluator class is very dependent on a Simulation class instance - not good, should be able to evaluate models with just a file/folder path for csv files.
SImulation class is also not very flexible when it comes to converting to a csv file - "disease_cases" is hard coded.
Same with plotting